Refactor skills trought a React.FC
