+++ pages/404.tsx
@@ -1,7 +1,7 @@
 import { NextPage } from 'next'
 import Link from 'next/link'
 
-import Layout from '~components/Layout';
+import Layout from '~src/components/Layout';
 
 const Custom404: NextPage = () => {
   return (
