+++ pages/github.tsx
@@ -1,6 +1,6 @@
 import type { NextPage } from 'next'
 
-import Layout from '~components/Layout';
+import Layout from '~src/components/Layout';
 
 const Github: NextPage = () => {
   return (
