+++ src/@types/skill.ts
@@ -1,0 +1,4 @@
+export interface SkillProps {
+    language: string;
+    progress: number;
+}
