+++ src/constants/skills.ts
@@ -1,0 +1,27 @@
+export const skills = [
+  {
+    language: 'Javascript',
+    progress: 80,
+  },
+  {
+    language: 'Java',
+    progress: 60,
+  },
+  {
+    language: 'PHP',
+    progress: 50,
+  },
+  {
+    language: 'Typescript',
+    progress: 98,
+  },
+  {
+    language: 'Gloang',
+    progress: 70,
+  },
+  {
+    language: 'NodeJS',
+    progress: 75,
+  },
+  
+]
